requirements.txt
crewai==0.28.8
crewai[tools]==0.28.8
langchain==0.1.20
langchain-openai==0.1.8
langchain-community==0.0.38
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
